+++ apps/nouns-camp/src/app/new/[[...segments]]/page.js
@@ -5,9 +5,6 @@ import { useWallet } from "@/hooks/wallet";
 import ClientAppProvider from "@/app/client-app-provider";
 import ProposalOrTopicEditorScreen from "@/components/proposal-or-topic-editor-screen";
 import ConnectWalletScreen from "@/components/connect-wallet-screen";
-import { useCollection as useDrafts } from "@/hooks/drafts";
-import { useDialog } from "@/hooks/global-dialogs";
-import { useNavigate, useSearchParams } from "@/hooks/navigation";
 
 export default function Page(props) {
   const params = use(props.params);
@@ -15,40 +12,11 @@ export default function Page(props) {
   return (
     <ClientAppProvider>
       <RequireConnectedAccount>
-        <PageContent draftId={draftId} />
+        <ProposalOrTopicEditorScreen draftId={draftId} />
       </RequireConnectedAccount>
     </ClientAppProvider>
   );
 }
-
-const PageContent = ({ draftId }) => {
-  const [searchParams] = useSearchParams();
-  const navigate = useNavigate();
-  const { createItem: createDraft } = useDrafts();
-
-  const { open: openTemplateDialog, isOpen: isTemplateDialogOpen } =
-    useDialog("proposal-templates");
-
-  React.useEffect(() => {
-    if (draftId == null && searchParams.get("topic") != null) {
-      const draft = createDraft({ actions: null });
-      navigate(`/new/${draft.id}`, { replace: true });
-      return;
-    }
-
-    if (draftId == null) openTemplateDialog();
-  }, [draftId, searchParams, createDraft, navigate, openTemplateDialog]);
-
-  React.useEffect(() => {
-    if (draftId == null && !isTemplateDialogOpen) {
-      navigate("/", { replace: true });
-    }
-  }, [draftId, isTemplateDialogOpen, navigate]);
-
-  if (draftId == null) return null;
-
-  return <ProposalOrTopicEditorScreen draftId={draftId} />;
-};
 
 const RequireConnectedAccount = ({ children }) => {
   const { address: connectedAccountAddress } = useWallet();